refactor: Popover 포지셔닝 로직 분리 [TM-127]
