+++ src/widgets/calendar/config/calendar.constants.ts
